gateway/requirements.txt
Flask==3.1.0
grpcio==1.68.1
grpcio-tools==1.68.1
pika==1.3.2
pymongo==4.10.1
redis==5.2.1
prometheus-client==0.21.1